+++ Pipfile
@@ -7,6 +7,7 @@ verify_ssl = true
 
 [packages]
 django = "==3.2"
+django-widget-tweaks = "*"
 
 [requires]
 python_version = "3.7"
